Move http service to separate folder
