+++ src/services/http/http.test.ts
@@ -1,8 +1,8 @@
 import 'isomorphic-fetch';
 import tape from 'tape';
 import sinon from 'sinon';
-import HttpService from './http';
-import { stubCall, stubFetch } from '../util/testUtils';
+import HttpService from '.';
+import { stubCall, stubFetch } from '../../util/testUtils';
 
 const post = {
   type: 'POST',
